requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
asyncmock==0.4.2
attrs==25.1.0
blessed==1.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
debugpy==1.8.12
distro==1.9.0
Django==4.2.19
django-environ==0.12.0
django-extensions==3.2.3
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.15.2
Faker==36.1.1
fastapi==0.115.8
frozenlist==1.5.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.3.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
lanarky==0.8.6
loguru==0.7.3
mock==5.2.0
multidict==6.1.0
openai==1.64.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
reactivex==4.0.4
redis==3.5.3
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==1.8.2
starlette==0.45.3
tiktoken==0.4.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.29.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==11.0.3
yarl==1.18.3
